refactor: structured components by common, features, layout, and ui. Created index.js files for cleaner imports.
