+++ client/src/components/common/buttons/FloatingCreateButton.jsx
@@ -1,5 +1,5 @@
 import { useNavigate } from "react-router-dom";
-import { supabase } from "../supabaseClient";
+import { supabase } from "../../../supabaseClient";
 import React, { useState, useEffect } from "react";
 import { toast } from "react-toastify";
