+++ client/src/components/common/buttons/index.js
@@ -1,0 +1,2 @@
+export { default as FloatingCreateButton } from "./FloatingCreateButton";
+export { default as HeartButton } from "./HeartButton";
