+++ client/src/components/common/modals/index.js
@@ -1,0 +1,1 @@
+export { default as HelpModal } from "./HelpModal";
