+++ client/src/components/common/navigation/StickyHeader.jsx
