+++ client/src/components/common/navigation/index.js
@@ -1,0 +1,1 @@
+export { default as StickyHeader } from "./StickyHeader";
