+++ client/src/components/features/auth/AuthWrapper.js
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { supabase } from "../supabaseClient";
+import { supabase } from "../../../supabaseClient";
 import { useNavigate } from "react-router-dom";
 
 const AuthWrapper = ({ children }) => {
