+++ client/src/components/features/auth/ProfileCheckWrapper.js
@@ -1,7 +1,7 @@
 // components/ProfileCheckWrapper.js
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { supabase } from "../supabaseClient";
+import { supabase } from "../../../supabaseClient";
 
 const ProfileCheckWrapper = ({ children }) => {
   const [loading, setLoading] = useState(true);
