+++ client/src/components/features/auth/index.js
@@ -1,0 +1,2 @@
+export { default as AuthWrapper } from "./AuthWrapper";
+export { default as ProfileCheckWrapper } from "./ProfileCheckWrapper";
