+++ client/src/components/features/map/index.js
@@ -1,0 +1,1 @@
+export { default as MapComponent } from "./MapComponent";
