+++ client/src/components/features/profile/TestDataInsert.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { supabase } from "../supabaseClient";
+import { supabase } from "../../../supabaseClient";
 
 const TestDataInsert = () => {
   const insertTestData = async () => {
