+++ client/src/components/features/profile/index.js
@@ -1,0 +1,1 @@
+export { default as TestDataInsert } from "./TestDataInsert";
