+++ client/src/components/features/restaurant/AllergenSummary.js
