+++ client/src/components/features/restaurant/index.js
@@ -1,0 +1,3 @@
+export { default as RestaurantCard } from "./RestaurantCard";
+export { default as ReviewCard } from "./ReviewCard";
+export { default as AllergenSummary } from "./AllergenSummary";
