+++ client/src/components/features/search/SearchBox.js
