+++ client/src/components/features/search/index.js
@@ -1,0 +1,1 @@
+export { default as SearchBox } from "./SearchBox";
