+++ client/src/components/index.js
@@ -1,0 +1,14 @@
+// Layout components
+export * from "./layout";
+
+// Feature components
+export * from "./features/auth";
+export * from "./features/restaurant";
+export * from "./features/map";
+export * from "./features/profile";
+export * from "./features/search";
+
+// Common components
+export * from "./common/buttons";
+export * from "./common/modals";
+export * from "./common/navigation";
