+++ client/src/pages/Explore.js
@@ -1,9 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { supabase } from "../supabaseClient";
-import MapComponent from "../components/MapComponent";
-import SearchBox from "../components/SearchBox";
-import RestaurantCard from "../components/RestaurantCard";
-import FloatingCreateButton from "../components/FloatingCreateButton";
+import { MapComponent, SearchBox, RestaurantCard, FloatingCreateButton } from "../components";
 
 const ALLERGENS = ["Peanut", "Tree Nut", "Gluten", "Wheat", "Eggs", "Milk", "Sesame", "Soy", "Fish", "Shellfish"];
 const RESTAURANT_TYPES = ["Bakery", "Ice Cream", "Boba", "Cafe", "Fast Food", "Diner", "Casual Dining", "Fine Dining", "Vegan", "Vegetarian", "Italian", "Seafood", "Pizza", "BBQ", "Mexican", "Indian", "Asian", "Mediterranean"];
